Move window manager exclusive files to manager module
