+++ src/common/mod.rs
@@ -5,15 +5,12 @@ mod monitor_handle;
 mod monitor_info;
 mod monitors;
 mod persistent_workspace_id;
-mod placement;
 mod point;
 mod rect;
-mod scrolling_strips;
 mod sizing;
 #[cfg(test)]
 mod tests;
 mod transient_workspace_id;
-mod width_preset;
 mod window;
 mod window_handle;
 mod window_placement;
@@ -27,13 +24,10 @@ pub use crate::common::monitor_handle::MonitorHandle;
 pub use crate::common::monitor_info::MonitorInfo;
 pub use crate::common::monitors::Monitors;
 pub use crate::common::persistent_workspace_id::PersistentWorkspaceId;
-pub(crate) use crate::common::placement::Placement;
 pub use crate::common::point::Point;
 pub use crate::common::rect::Rect;
-pub(crate) use crate::common::scrolling_strips::ScrollingStrips;
 pub use crate::common::sizing::Sizing;
 pub use crate::common::transient_workspace_id::TransientWorkspaceId;
-pub(crate) use crate::common::width_preset::WidthPreset;
 pub use crate::common::window::Window;
 pub use crate::common::window_handle::WindowHandle;
 pub use crate::common::window_placement::WindowPlacement;
