+++ src/common/tests/mod.rs
@@ -1,4 +1,2 @@
-mod placement_tests;
-mod scrolling_strips_tests;
 mod sizing_tests;
 mod workspace_tests;
