+++ src/window_manager/mod.rs
@@ -1,8 +1,11 @@
 mod navigation;
+mod placement;
 mod scrolling_layout;
+mod scrolling_strips;
 mod spatial_layout;
 #[cfg(test)]
 mod tests;
+mod width_preset;
 #[allow(clippy::module_inception)]
 mod window_manager;
