+++ src/window_manager/tests/scrolling_strips_tests.rs
@@ -1,4 +1,6 @@
-use crate::common::{Direction, PersistentWorkspaceId, Rect, ScrollingStrips, Sizing, WidthPreset, WindowHandle};
+use crate::common::{Direction, PersistentWorkspaceId, Rect, Sizing, WindowHandle};
+use crate::window_manager::scrolling_strips::ScrollingStrips;
+use crate::window_manager::width_preset::WidthPreset;
 
 fn workspace(number: usize) -> PersistentWorkspaceId {
   PersistentWorkspaceId::new([number as u16; 32], number, number == 1)
