+++ src/window_manager/width_preset.rs
